// src/app/interfaces/etiqueta.ts
export interface Etiqueta {
    id_etiqueta?: number;
    etiqueta?: string;
    nombre?: string;
    descripcion?: string;
    fecha?: string;
    fotografia?: boolean;
    icono?: number;
    nombre_icono?: string;
    descripcion_icono?: string;
    fecha_edicion?: string;
    color_icono?: string;
}